Fix issue with invalid count value

+++ service/pictures.go
@@ -12,7 +12,7 @@ import (
 type PicturesService interface {
 	Create(*multipart.FileHeader) (*dto.PictureResponse, *dto.InvalidPictureFileError)
 	Update(int, *multipart.FileHeader) (*dto.PictureResponse, *dto.InvalidPictureFileError)
-	List(int, int) ([]*dto.PictureResponse, int, error)
+	List(int, int) ([]*dto.PictureResponse, int64, error)
 	Get(int) (*dto.PictureResponse, error)
 	GetFile(int) (string, error)
 	Delete(int) error
@@ -61,7 +61,7 @@ func (s *picturesService) Update(id int, file *multipart.FileHeader) (*dto.Pictu
 	return picture.ToPictureResponse(), nil
 }
 
-func (s *picturesService) List(limit, page int) ([]*dto.PictureResponse, int, error) {
+func (s *picturesService) List(limit, page int) ([]*dto.PictureResponse, int64, error) {
 	pictures, totalCount, err := s.repository.GetAll(limit, page)
 	if err != nil {
 		return nil, 0, err
